feat: weekly SUB changes Reddit draft on Mondays
Extract the card-wire fetch/collapse logic from post-weekly-sub-changes.js
into scripts/lib/weekly-sub-changes.js and add a Reddit counterpart that
renders the full week of net changes (no tweet-length truncation) and sends
it to the Social Posting Service with platforms: ["reddit"] + publish_now.
The service's manual reddit platform records a pre-filled submit URL as
pending_manual; nothing auto-publishes to Reddit. New Monday 15:00 UTC
workflow mirrors weekly-sub-changes.yml.

The body is deliberately plain text (no markdown tables/bold): Reddit's
submit page opens in the rich-text editor, which takes pre-filled text
literally, so markdown syntax would post as visible asterisks and pipes.
Verified against the live composer.
