+++ .github/workflows/weekly-sub-changes-reddit.yml
@@ -1,0 +1,26 @@
+name: Weekly Sign-up Bonus Changes Reddit Draft
+
+on:
+  schedule:
+    # Every Monday at 3pm UTC (11am ET during DST / 10am ET otherwise).
+    # Nothing auto-publishes: this drops a pending-manual Reddit draft into
+    # the Social Posting Service UI, so blackout windows don't apply.
+    - cron: '0 15 * * 1'
+  workflow_dispatch: {}
+
+jobs:
+  draft-reddit-sub-changes:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v6
+
+      - uses: actions/setup-node@v6
+        with:
+          node-version: 18
+
+      # No npm ci: the script only uses Node built-ins and repo-local requires.
+      - name: Create weekly sign-up bonus changes Reddit draft
+        env:
+          SOCIAL_API_URL: ${{ secrets.SOCIAL_API_URL }}
+          SOCIAL_API_KEY: ${{ secrets.SOCIAL_API_KEY }}
+        run: node scripts/post-weekly-sub-changes-reddit.js
